refkit image: flexible partition sizes
The VFAT parition size specified in the DSK_IMAGE_LAYOUT no longer
matched the actual size used for images, and 15MB indeed is barely
large enough. Adding tools like cryptsetup to the initramfs causes it
to exceed that limit.

Instead of hard-coding a fixed size in the .wks.in file, now a
per-image variable is inserted instead. The size of the rootfs can
also be controlled, including not limiting it at all and thus adapting
the image size to the size of the rootfs. This will be needed for the
installer image which is larger than normal images.

The default is the same as before (30MB for the VFAT partitions, 4GB
total image size).

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>