refkit + linux-yocto/intel: enable dm-verity
This enables dm-verity in all kernels called "linux-yocto" or "linux-intel"
if the "dm-verity" distro feature is selected. This approach is slightly
easier than a linux-yocto bbappend; adding support for kernels
with other names can be added with a one-line change.

Hosting the kernel config fragment locally is an interim solution
until a pending patch for yocto-kernel-patch gets accepted and is
available.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>

merge: enable dm-verity in kernel