meta-iotqa: Ignore Minnowboard boot error
Add '*ERROR* dp aux hw did not signal timeout' to list of ignored
errors. This error happens randomly under 5% of the time and doesn't
seem to cause any additional errors in other tests. Tried to reproduce
the error with an image that it happened with but couldn't get it with 25
flash/test cycles. This error also doesn't seem to be easily fixed. So ignore this
error as it will cause tests to fail randomly.

Signed-off-by: Simo Kuusela <simo.kuusela@intel.com>