refkit.conf: enable dm-verity and luks
The result is that the installer image will be read-only with
dm-verity as integrity protection. The main practical advantage is
that bit flips on low-quality USB sticks will be detected reliably.

Whole disk-encryption with LUKS is enabled for normal images, but not
enforced and thus does not really protect against offline attacks. It
gets activated for images installed by the installer image.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>