meta-iotqa: Add mraa GPIO test
The test will set a GPIO pin on and off using mraa and check the output
to confirm it works. The test requires to use BeagleBone Black as the
testing host and setting up it's GPIO pin 20 as input. Also correct
cabling is required for checking the clients GPIO pin output.

Signed-off-by: Simo Kuusela <simo.kuusela@intel.com>