systemd-boot: use built-in RMC db and fix cmdline mangling
To address problems in RMC usage together with secure boot,
change the stub such that the RMC db is used only if it's
built in the executable.

Fixes [YOCTO 11030/partial].

Some UEFI shells (seen with Intel NUCs) indicate LoadOptionSize
is > 0 but there's actually no data (*LoadOptions == 0). This
wipes the build in command line and we loose, e.g., root=PARTUUID
setting and the boards fail to boot. Add a check that peeks if
there's data in *LoadOptions before mangling the cmdline.

Fixes [YOCTO 11078].

Signed-off-by: Mikko Ylinen <mikko.ylinen@linux.intel.com>