refkit-ci.inc: CI specific HOSTTOOLS setup
OE-Core recently added a whitelist check against what host
tools are OK to be used and a check that they also exist in
PATH.

The list is specified by HOSTTOOLS variable. The default value
added host tools that are new (i.e., not installed) in the CI
configuration.

This commit customizes the default HOSTTOOLS to match with CI
distro configuration.

Signed-off-by: Mikko Ylinen <mikko.ylinen@linux.intel.com>