meta-iotqa: extend runtest.py to accept target's port number
In case of testing a QEMU-based target it might be easier to
use SLIRP networking which doesn't require root permissions to
create a TAP network interface. But it means that the target can
be accessible only via a port forwarded to the host's loopback
interface. And as the default SSH port is usually taken already
the target's SSH port is forwarded to port 2222 on localhost.

Unfortunately the runtest.py script assumes that the SSH
port is 22 always.

This patch extends runtest.py options to accept an optional
port argument. So the command to run tests on a QEMU target
may look like the following:

    runtest.py -f testplan/refkit-image-common.manifest -m intel-corei7-64 -t 127.0.0.1:2222 -s 10.0.2.2

Signed-off-by: Dmitry Rozhkov <dmitry.rozhkov@linux.intel.com>