refkit-image.bbclass: remove obsolete tweaks for "login" and "passwd"
"shadow" gets installed also without the IMAGE_INSTALL change, perhaps
because it now provides the "login" command. Busybox no longer has
that.

Using shadow passwd is probably still relevant in a stateless
configuration which uses nss-altfiles, but not in the current refkit
images. If it's needed, the tweak needs to go into the stateless
configuration file.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>