refkit-images: make swupd-related changes optional
Several of the changes in refkit-image.bbclass were unconditionally
enabled although they had been added to support the swupd feature.
The thinking was that images with and without swupd should be as
similar as possible.

However, swupd is not currently officially supported, so this kind
of special treatment is no longer justified. Instead, these additional
settings are now in the separate refkit-swupd-image.bbclass, which
only gets inherited when the swupd image feature is active.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>