supported recipes: support multiconfig, add separate script
The supported recipes report can now be generated in two way: as
before with SUPPORTED_RECIPES_SOURCES in supported-recipes.bbclass,
and with a dedicated supported-recipes-report.py script.

The latter is easier to use, but probably needs a bitbake more recent
than Yocto 2.2. The supported-recipes.bbclass is meant to continue
working also with older Yocto releases.

At the moment, the script still relies on the build configuration
already inheriting supported-recipes.bbclass. Perhaps that limitation
can be removed. In contrast to the bbclass, the script does not rely
on additional dump files with information about the sources used by a
recipe. Instead, it parses each active recipe on demand to retrieve
the information. This is a bit slower, but cleaner.

Both approaches now support multiconfig. Shared sources are only
included once in the report.

"-" can be used to write the report to stdout.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>