initramfs-framework-refkit-luks.bb: optionally require encryption
Refkit images need to work without encryption (booting from USB stick)
and with encryption (installed to internal disk), and therefore the
LUKS initramfs code checks whether a partition is using LUKS and
continues without LUKS if it is not.

This can be changed by building an image with "use_encryption" as boot
parameter in the APPEND image variable. However, such an image then
can only be used after installing it to internal disk.

Long-term it would be better to set the additional boot parameter
during installation, instead of having to build it into the image
itself.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>