supported-recipes.bbclass: ignore recipes without packages
Several recipes which previously had to be listed explicitly inherit
nopackages.bbclass, so we can use that inheritance to exclude them
instead.

Recipes which do not generate packages are very unlikely to add
content to an image for the target (technically it would be possible
in an image recipe, but those already get excluded via the
image.bbclass base class anyway).

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>