WiP: refkit-sanity.bbclass: try teaching about OSTree.
Sigh...

Try teaching refkit_qa_image about the peculiarities of OSTree-based
images. Again... this might not be the right approach. The easiest
would be to simply let the qa check check the pristine rootfs we use
as input and trust our conversions+boot-time-gimmickery to be correct
and not break the image.