image-installer selftest: log bitbake build invocation
Starting the first test can take a while because it needs to build the
images. The developers should get a notice about that. We also see
unexpected rebuilds in the CI, so also include what bitbake was actually
doing once it is done.

Live logging of the output would be even nicer, but would depend on
enhancements in meta/oeqa/utils/commands.py.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>