ostree: initial integration into refkit
The OSTree recipe is taken from meta-flatpak, because that is what is
already in intel-iot-refkit. The integration into the build process
relies on using wic. Special support is provided for updating the UEFI
combo app.

This commit combines the initial work done by Krisztian Litkey with
various cleanup changes by Patrick Ohly. It got squashed because the
entire history would be confusing.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>