meta-refkit: Add test to check qemu
Add a basic test to check that qemu runs with refkit. Based on
Yocto bug #11203. The test is to be add in a selftest/refkit
path, so it could be run with:

oe-selftest -r refkit.runqemu.RunqemuTests.test_boot_machine

Further selftest tests can be added in the refkit folder. Considering
this, running all refkit selftest tests could be easily done with:

oe-selftest --run-tests-by module refkit.*

Signed-off-by: Humberto Ibarra <humberto.ibarra.lopez@intel.com>