refkit-image.bbclass: use virtio under runqemu
The default is to use SCSI. virtio support is enabled by our upstream
kernel configuration and provides a performance boost during testing,
so let's use it.

A better place for this might be in the machine configuration, but
it's not there at the moment and also can't be set there because of
the ?= default in qemuboot.bbclass. A proposal to change that was sent
to the OE-core mailing list ("Re: [OE-core] [PATCH 4/7] runqemu: do
not rely on grepping images").

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>