sbsigntool: add secure boot signing tool
sbsigntool is a signing tool used by, e.g., Ubuntu, to sign EFI
applications to be able to boot the OS on a secure boot enabled system.

In Refkit, sbsigntool is added in the build process and used to sign
the UEFI combo application if the image needs to have secure boot compatible
EFI app.

Signed-off-by: Mikko Ylinen <mikko.ylinen@linux.intel.com>