Enable QEMU testing
For QEMU testing qemu.mask file has been created that includes all tests
that can't be run on QEMU. Also if-else structure has been added to
tester-exec.sh so tests on real devices are run using DAFT and QEMU
tests are run with some bash and expect script. The expect script is
used to run the QEMU and add an SSH key to it so tests can be run.

Signed-off-by: Simo Kuusela <simo.kuusela@intel.com>