CI Jenkinsfile: stop calculating testinfo sumsize
Summed size of testinfo was used to detect need
for tester jobs, but same result is achieved
by checking size of testinfo in for-loop, eliminating
need for extra variable.
Later, same variable was used to gate report sending,
but report should be sent even if no tests were run,
as it now contains also other information.

Signed-off-by: Olev Kartau <olev.kartau@intel.com>