meta-refkit-extra: build testing for computer vision demo
Our CI system does not enable the layer by default to avoid tainting
"normal" builds. However, we need to ensure that the extra content at
least continues to build.

This can be done in a meta-refkit selftest which adds the extra layer
as needed and removes it again. This is sub-optimal because it adds a
long-running test with no output while the test runs; extending the CI
setup so that it covers different build configurations would be
better. But for now the selftest is better than nothing.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>