image-dsk.bbclass: rebuilt when Secure Boot key files change
This replaces the dependency on the file names with a dependency on
the file content, i.e. it becomes possible to move files around
without a rebuild and a rebuilt gets properly triggered when
generating new keys under the same name.

This a straight copy of the approach used for dm-verity in
refkit-hash-dm-verity-key.bbclass.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>