intel-iot-refkit: move content into meta-refkit-core/computervision/gateway layers
The intention is to make it simpler for other projects to use content
from intel-iot-refkit without having to use the entire distro. This
first commit is good enough to keep building the refkit distro and
just moves files around without modifying them. Further work is needed
to make the individual pieces also functional stand-alone.

Separate profile layers get introduced and define their dependencies,
because then existing tooling (layer index, Toaster,
yocto-compat-layer.py) will be able to set up a project using them
together with these dependencies.

The meta-refkit-core layer is meant to be more flexible and does not
enforce the use of any layers besides OE-core. At the moment there are
probably still parse errors (due to .bbappends) or build errors
(missing dependencies) when layers are missing. This will be addressed
separately on a case-by-case basis.

The meta-refkit distro layer is meant to have just a minimal distro
configuration file which includes files from the other layers, plus
the project-specific content:
- local.conf and bblayers.conf samples
- the "supported recipes" list, which gets maintained for the distro and
  not individually for each layer
- selftests which depend on the exact project configuration

This new approach gets documented in the doc/introduction.rst file, in
addition to explaining a bit more the benefit of using the
intel-iot-refkit and its distro as-is.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>