refkit-init-build-env: wrap it around oe-init-build-env
oe-init-build-env allows to use a wrapper script that just
sets the necessary (custom) values to operate with.

Since oe-init-build-env does most of the init work, it's not needed to
maintain the same in refkit-init-build-env. Therefore, simply make
refkit-init-build-env a wrapper around oe-init-build-env. This makes
refkit-init-build-env simple.

Furthermore, this change fixes a problem where it was not possible to source
the build-env script outside the meta-data dir (intel-iot-refkit). With the
fix also it's possible to have the build directory separate from the
meta-data dir.

Fixes: [YOCTO #11402]

Signed-off-by: Mikko Ylinen <mikko.ylinen@linux.intel.com>