licensecheck: library for finding package runtime licenses.
Licensecheck calculates a given package's runtime dependency tree. It
then uses that information to propagate licenses from bottom towares the
tree root, eliminating licenses if they are inconsistent or not allowed
to be used.

Note that for many reasons (dual- and multi-licensing, different source
licenses for a package, ...), and also for difficulties in putting to a
simple table the relationships between licenses, this library does in no
way guarantee that the licensing of a component is correct, even if the
library doesn't report any problems. The library is just a tool for
finding possible inconsistencies in licensing.

Signed-off-by: Ismo Puustinen <ismo.puustinen@intel.com>