selftests: check if runtime licensing of an image is valid.
The refkit-license-check.py selftest is intended for finding out if
there are licensing conflicts in the images or if prohibited licenses
are used in installed components. It uses licensecheck library from
meta-refkit.

Currently one test is implemented. The test tries to find out whether or
not there are software components used under (L)GPLv3 license in
refkit-image-computervision production configuration.

Note that the test is not accurate and complex enough to be used as a
guarantee that the image licensing is valid and it doesn't contain any
(L)GPLv3 components. It is just a tool helping to catch some possible
mistakes.

Signed-off-by: Ismo Puustinen <ismo.puustinen@intel.com>