dm-verity: require key if and only if feature enabled
This makes it possible to easier to build images without the feature
because in that case, there is no need to provide a key.

initramfs-framework-refkit-dm-verity.bb makes no sense without a key
and now removes itself from a world build when no key is set. Trying
to use it will then trigger a dependency error.

refkit-initramfs.bb cannot check whether the
initramfs-framework-refkit-dm-verity is available and thus continues
to check the DISTRO_FEATURES to determine whether it should enable
dm-verity.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>