refkit-installer-image: fix install to eMMCs
The $output*$gdisk_pnum wildcard listing is too greedy
when the output is an eMMC device: it matches not only
'p' (in mmcblkXpY) but also 'boot' which is also found
on eMMC devices. This results in a 'partition is not
unique' error.

To fix the install, limit the search to <device><partnum>
and <device>p<partnum>.

Signed-off-by: Mikko Ylinen <mikko.ylinen@linux.intel.com>