refkit-poky.py: detect broken images
yocto-compat-layer.py did not detect missing dependencies in
meta-refkit-gateway because the refkit-image-gateway was excluded from
world builds. We need to explicitly override that default. However,
this only works for images added by the layer that is getting tested,
because existing images change their signature (unnecessarily?) when
BBLAYERS gets changed.

For this config change to work with self.append_config(),
test_compat_* tests must use the same scaffolding as the rest of the
test, so we might as well move that to setUp/tearDownLocal().

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>