scripts: added script for generating a flatpak repostory.
Added a workhorse script for generating/populating a primary flatpak
repository for an image. The script can also replicate this repository
to another archive-z2 repository which can then be exposed over HTTP
for consumption by flatpak clients. This will be necessary for building
flatpaks for/against a flatpak-enabled refkit image.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>