meta-refkit-core: Use separate bbappends directory
The new "bbappends" hierarchy can be used to provide .bbappends for
recipes in layers which may or may not be present. This relies on the
BBFILES_DYNAMIC support in bitbake master.

The complete hierarchy for all layers that are currently used by the
refkit distro is set up in advance with empty directories where
.bbappends need to go, to make it easier for developers to add
.bbappends correctly.

.bbappends outside of the "bbappends" hierarchy are ignored.

While technically not required because OE-core is a hard dependency,
also hosting the .bbappends that modified OE-core under the new
"bbappends" hierarchy makes more sense (all modifications to other
layers in one place).

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>