gnupg1: added native-only gnupg1 for key generation/repo signing.
Added a native-only recipe for gnupg1. We'll use this to generate
keys and sign (ostree) repository commits during bitbake/image
building in connection with ostree and flatpak repositories.

While gnupg2 can be beaten to working obedience with hacks, it is
somewhat of a PITA to use both in non-interactive batch mode and
in the Yocto relocated native environment. Instead of dealing with
those problems we generate/import signing keys and sign repository
commits using gnupg1, and check signatures on client devices using
gnupg2.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>