selftest: secureboot: don't add 'serial' with runqemu
runqemu in oe-core recently changed _some_ subprocess
module calls to use check_call() to better catch failures.

This hits us because the 'stty' check_call()s in runqemu
via oe-selftests on Docker (in CI) raise error due to:

stty: 'standard input': Inappropriate ioctl for device

It seems 'stty' does work with Docker's pseudo-tty
implementation.

There are other 'stty' calls in runqemu too but they don't
use subprocess.check_call() or are not run due to conditional
blocks so we're not bit by those.

The failing 'stty' is run if 'serial' runqemu parameter is used.
So we just drop that and we're OK (and feeling lucky).

Signed-off-by: Mikko Ylinen <mikko.ylinen@linux.intel.com>