flatpak-*.bbclass: added flatpak image and repository support classes.
Added three classes for supporting building flatpak-enabled images, and
creating/publishing repositories with flatpak platform and SDK runtimes
for such images.

flatpak-config.bbclass contains the flatpak-related bitbake/configuration
variables. It is separated to a class of its own because we'll probably
need to inherit it in a recipe which will support pre-populating an image
with pre-declared flatpak application repositories.

flatpak-image-variants.bbclass declares two image variants: flatpak-
runtime and flatpak-sdk. flatpak-runtime is our flatpak-enabled runtime
image variant. It is supposed to get flashed to client devices that need
flatpak support. flatpak-sdk is a corresponding image which we use to
populate a flatpak repository with a flatpak SDK runtime which can then
be used to compile/generate flatpaks for the flatpak-runtime variant.

flatpak-repository.bbclass takes care of the details of populating and
exporting flatpak repositories with images.

Signed-off-by: Krisztian Litkey <kli@iki.fi>