meta-refkit-industrial: switch ROS to TinyXML2
since it's maintained better than deprecated TinyXML.

The patch set is not going to be merged to the upstream ROS
layer because only ROS Indigo release is maintained in the layer
currently.

In https://github.com/ros/rospack/pull/62 it was suggested that
the packages in the ROS Indigo release are not going to switch
to TinyXML2, but it may happen in future releases.
Also a comment in https://github.com/ros/urdfdom_headers/pull/35
suggests that the switch is precluded by the fact TinyXML types
are widely used in urdfdom's public API thus requires fixing all
downstream packages which can be difficult if possible at all.

Signed-off-by: Dmitry Rozhkov <dmitry.rozhkov@linux.intel.com>