refkit poky testing: reduce testing of refkit-image-common
We cannot guarantee that all layers actually work with OE-core master
and its Poky distro. For example, when OE-core switched to gcc7,
nodejs and upm failed to build.

Therefore we now limit testing in that build configuration to OE-core
and meta-refkit-core, without any of the additional layers.

As an additional benefit, the test now ensures that meta-refkit-core
is actually usable for image building and booting when just using
OE-core.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>