docker: install netcat in OpenSUSE container
The new OSTree testing needs netcat to forward HTTP requests from the
guest to the host. Adding more dependencies on the host system is not
nice, but a lot easier than staging native tools with bitbake.

Connecting guest to host without relying on the tool was also tried,
but then qemu keeps the TCP connection to the HTTP server open all the
time, so it is not quite the same and didn't work in practice.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>