python(3): set ALTERNATIVEs for target packages only
buildtools-tarball used by the eSDK installs nativesdk-python(3)*
and those are used instead of host's python when the eSDK is installed
and run.

When 'python3' is installed/linked to 'python', the eSDK setup errors
with:

"OpenEmbedded requires 'python' to be python v2 (>= 2.7.3), not python v3.
 Please set up python v2 as your default 'python' interpreter."

To fix the install problem, set python/python_config ALTERNATIVEs for
target packages only.

Signed-off-by: Mikko Ylinen <mikko.ylinen@linux.intel.com>