flatpak-repository.bbclass: added file based runtime type detection.
If not obvious from the image name go through the image content
and check for the presence of a few files to determine if the
image is a BasePlatform or a BaseSdk flatpak runtime.

This is now necessary since/if we make the basic gateway image
flatpak enabled and also want to generate an accompanying flatpak
repository for it.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>