selftest: log command output
With the recently enhanced runCmd/bitbake() in OE-core master we can
get output from the invoked command passed through the logger that
each test has, which provides much nicer output while a test runs:
instead of test that just seems to hang, one gets a constant stream
of messages with time stamps.

The command also gets logged, so we can remove some extra output which
now is a bit redundant.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>