introduction.rst: add a note about eSDKs
The extensible SDK (eSDK) is an important piece of
the overall development story. Therefore, a note about its
existence makes sense in the Reference Kit introduction.rst too.

To get started with the eSDK in refkit, we primarily rely
on the comprehensive Yocto Project SDK manual. The SDK manual
provides the necessary information how to build the eSDK
installer for refkit profiles/images.

Completes [YOCTO: #11493].

Signed-off-by: Mikko Ylinen <mikko.ylinen@linux.intel.com>