system update: test user handling
The main part of the test is that a new system user is active after an
update. Optionally, the test does the update while a local user has
been added.

The normal OSTree /etc handling is not enough for the full test: once
/etc/passwd and /etc/group were modified by adding the local user,
that copy of the files continue to be used and the new system user is
not registered after the update.

So for now we only enable the "light" test.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>