system update: test with dedicated image recipes
Using "refkit-image-update-ostree.bb" instead of the normal
"refkit-image-common.bb" allows testing without affecting normal use
and/or other tests.

Partly fixes: [YOCTO #11757]

With OSTree it is fairly easy to build the original image and then the
update in a different image recipes. This makes it easier to debug the
test, because both build directories are preserved. It also speeds up
running the tests repeatedly, because the rootfs or even the actual
image files can get reused. Because this is not how OSTree is normally
used, the normal testing as still using a single
recipe. refkit_ostree.RefkitOSTreeUpdateTestDev is for local testing
by a developer.

In the future, this base infrastructure could be used to test
switching a device between different images with OSTree.

While at it, build the update image later. That makes it easier
to debug problems that already occur after building and booting the
original image.

Because bitbake deletes obsolete build artifacts, the image recipes
have to be permanently available. Creating them on-the-fly does not
work because their WORKDIR then gets removed when bitbake gets invoked
while the recipe is gone (see
https://bugzilla.yoctoproject.org/show_bug.cgi?id=11771).

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>