stateless: core mechanism + individual .inc files
This is a cleaned up and updated version of the stateless work that
was done before.

It introduces "stateless" as distro and image feature. All of the
functionality is in stateless.bbclass, which also has
documentation. Currently the class is meant to be inherited globally,
because it enables additional patches for individual recipes
according to the STATELESS_SRC variables set by the .inc files.

This approach was chosen as an interim solution because it keeps those
changes all in individual files, instead of scattering them across
different .bbappends or upstream recipes. Also, most of the patches
come straight from the Clear Linux Project without modifications, so
it cannot be guaranteed that they always apply. Ideally, the patches
should go into the recipes that they are patching or (better)
upstream.

Most of the functionality is around the rootfs transformation. The
idea here is that packages and rootfs creation do not need to be
modified in several cases. For images that use a whole-disk update
mechanism, additional post-processing can move configuration files
around and change them as needed.

In other cases (like not installing unnecessary example config files
into /etc), the upstream recipe would need to be changed, so there is
further work left for OE and Yocto. This commit can serve as a
starting point for a discussion arounnd that.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>